% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{almost_equal}
\alias{almost_equal}
\title{Compare Two Values Allowing a Tolerance}
\usage{
almost_equal(x, y, tolerance = 1e-06)
}
\arguments{
\item{x}{vector of numeric}

\item{y}{vector of numeric}

\item{tolerance}{tolerance, default: \code{1e-6}}
}
\value{
vector of logical
}
\description{
Compare Two Values Allowing a Tolerance
}
