% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_comma}
\alias{format_comma}
\title{turn decimal number into string with comma}
\usage{
format_comma(x, nsmall = 0, ...)
}
\arguments{
\item{x}{numeric vector}

\item{nsmall}{number of digits to show}

\item{...}{others}
}
\description{
turn decimal number into string with comma
}
