% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.R
\name{my_theme}
\alias{my_theme}
\title{turn decimal number into string with comma}
\usage{
my_theme(...)
}
\arguments{
\item{...}{others arguments passed to 'ggplot2::theme()'}
}
\description{
turn decimal number into string with comma
}
