% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bwb_colours.R
\name{plot_bwb_colours}
\alias{plot_bwb_colours}
\title{Plot the BWB Colours in a Barplot}
\usage{
plot_bwb_colours(simple = TRUE)
}
\arguments{
\item{simple}{passed to \code{\link{get_bwb_colours}}}
}
\description{
Plot the BWB Colours in a Barplot
}
