% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_csv}
\alias{read_csv}
\title{read csv file, e.g. variable rehab strategy or simulation results to compare}
\usage{
read_csv(file = NULL, dec)
}
\arguments{
\item{file}{path to csv file}

\item{dec}{decima semperator}
}
\description{
read csv file, e.g. variable rehab strategy or simulation results to compare
}
