% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_remove_all}
\alias{str_remove_all}
\title{Remove All Substrings Matching the Pattern}
\usage{
str_remove_all(x, pattern)
}
\arguments{
\item{x}{vector of character}

\item{pattern}{regular expression agains which the strings in \code{x} are
matched}
}
\value{
\code{x} with elements in which substrings matching the
  \code{pattern} are removed
}
\description{
Remove All Substrings Matching the Pattern
}
