% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{translate_data}
\alias{translate_data}
\title{Translate column names and categorical data}
\usage{
translate_data(df, translation_list)
}
\arguments{
\item{df}{data to be translated}

\item{translation_list}{list of names to be translated in format
`list(oldname1 = "newname1", oldname2 = "newname2", ...)`}
}
\description{
Translate column names and categorical data
}
