% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_sort_char}
\alias{unique_sort_char}
\title{Sorted Unique Character Values}
\usage{
unique_sort_char(x)
}
\arguments{
\item{x}{vector of objects that can be converted to a vector of character
with \code{as.character}}
}
\description{
Sorted Unique Character Values
}
