% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_spec_chars}
\alias{use_spec_chars}
\title{Replace Placeholders with Special Characters}
\usage{
use_spec_chars(x)
}
\arguments{
\item{x}{vector of character}
}
\description{
Replace Placeholders with Special Characters
}
