% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_csv}
\alias{write_csv}
\title{write csv file, e.g. simulation result table}
\usage{
write_csv(data, outdir, filename, dec)
}
\arguments{
\item{data}{dataframe to be written}

\item{outdir}{directory to save data}

\item{filename}{filename to save data}

\item{dec}{decima semperator}
}
\description{
write csv file, e.g. simulation result table
}
