% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_data_frame.R
\name{matrix_to_data_frame}
\alias{matrix_to_data_frame}
\title{Convert a Matrix to a Data Frame}
\usage{
matrix_to_data_frame(
  x,
  name_row = "row",
  name_column = "column",
  name_value = "value"
)
}
\arguments{
\item{x}{matrix}

\item{name_row}{name to be given to the data frame column containing the
row "coordinates". Default: \code{row}}

\item{name_column}{name to be given to the data frame column containing the
column "coordinates". Default: \code{column}}

\item{name_value}{name to be given to the data frame column containing the
matrix values. Default: \code{value}}
}
\value{
data frame with three columns: 1. row "coordinate", 2. column
  "coordinate", 3. value
}
\description{
Convert a Matrix to a Data Frame
}
\examples{
matrix_to_data_frame(matrix(1:12, nrow = 3, dimnames = list(1:3, 1:4)))

}
